requirements.txt
Flask==1.1.2
google-cloud-storage==1.36.1
gunicorn==20.0.4; python_version > '3.0'
Numpy==1.19.5
Pillow==8.0.1
xmltodict==0.12.0
opencv-python==4.4.0.46
